refactor: update routes in files
