+++ src/pages/home/index.tsx
@@ -1,7 +1,7 @@
 import "./home.scss";
 import React, { FunctionComponent, useState, useEffect, BaseSyntheticEvent } from "react";
 import HomeGrid from "./home-grid";
-import data from "../../data/data.json";
+import data from "../../utils/data.json";
 import { ICharacter } from "../../interfaces";
 
 const Home: FunctionComponent = () => {
